+++ Cargo.toml
@@ -4,7 +4,7 @@ members = [
     "aopt",
     "aopt-help",
     "cote",
-    "cote-derive-refactor",
+    "cote-derive",
     "simple-find-file",
     "snowball-follow",
 ]
