+++ cote-derive/src/gen/cote.rs
