+++ cote-derive/src/lib.rs
@@ -1,8 +1,10 @@
-mod analyzer;
-mod global;
+mod config;
+mod gen;
 mod value;
 
+use gen::Analyzer;
 use proc_macro_error::abort;
+use quote::quote;
 use syn::parse_macro_input;
 use syn::DeriveInput;
 
@@ -10,12 +12,23 @@ use syn::DeriveInput;
 #[proc_macro_error::proc_macro_error]
 pub fn parser(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let input: DeriveInput = parse_macro_input!(input);
+    let analyzer = Analyzer::new(&input).unwrap_or_else(|e| {
+        abort! {
+            input,
+            "Failed to analyzer `cote` configuration on struct: {:?}",
+            e
+        }
+    });
+    let impl_for_struct = analyzer.gen_all().unwrap_or_else(|e| {
+        abort! {
+            input,
+            "Failed to generate code for struct: {:?}",
+            e
+        }
+    });
 
-    analyzer::derive_parser(&input)
-        .unwrap_or_else(|e| {
-            abort! {
-                input, "Cote macro generate failed: {:?}", e
-            }
-        })
-        .into()
+    quote! {
+        #impl_for_struct
+    }
+    .into()
 }
